In some circumstances, the IDE (for example IntelliJ Idea) can fail when opening an existing project. In order to fix this, the container has to be run in a privileged mode.
Since Ubuntu 22.04, Firefox isn't anymore available as a native package
but as a snap one. In order to avoid to use snap within a Docker
container, a PPA repository providing Firefox is added.

Improve a little bit the scripts build.sh and run.sh. It is now possible
to give a custome name to the Docker image instead of
silverpeas/silverdev.